updated to have model groups and files
