+++ FourSquare-GroupProject/NetworkServices/FourSquareAPIClient.swift
@@ -9,50 +9,6 @@
 import Foundation
 import NetworkHelper
 
-struct Data: Codable & Equatable {
-    let response: Response
-}
-
-struct Response: Codable & Equatable {
-    let venues: [Venue]
-}
-
-struct Venue: Codable & Equatable {
-    let id: String
-    let name: String
-    let location: Location
-    let categories: [Category]
-    let hasPerk: Bool
-    let delivery: Delivery?
-}
-
-struct Location: Codable & Equatable {
-    let address: String
-    let lat: Double
-    let lng: Double
-    let postalCode: String
-    let city: String
-    let state: String
-    let country: String
-    let formattedAddress: [String]
-    let neighborhood: String?
-    let crossStreet: String?
-}
-struct Category: Codable & Equatable {
-    let shortName: String
-}
-
-struct Delivery: Codable & Equatable {
-    let id: String
-    let url: String
-    let provider: Provider
-}
-
-struct Provider: Codable & Equatable{
-    let name: String
-    //let icon: ProviderIcon
-}
-
 struct FourSquareAPIClient {
     static func getVenues(for near: String, query: String, completion: @escaping (Result<[Venue], AppError>) ->()) {
