+++ FourSquare-GroupProject/NetworkServices/Photo.swift
@@ -1,0 +1,9 @@
+//
+//  Photo.swift
+//  FourSquare-GroupProject
+//
+//  Created by Juan Ceballos on 2/24/20.
+//  Copyright © 2020 Juan Ceballos. All rights reserved.
+//
+
+import Foundation
